+++ test/helpers/poseidon-hash.ts
@@ -1,19 +1,18 @@
-import { ethers } from "hardhat";
+import { BaseContract, ContractFactory, hexlify, toBeHex } from "ethers";
 
-import { BaseContract } from "ethers";
+import { HardhatEthers } from "@nomicfoundation/hardhat-ethers/types";
 
+import { Poseidon } from "@iden3/js-crypto";
 // @ts-ignore
 import { poseidonContract } from "circomlibjs";
 
-import { Poseidon } from "@iden3/js-crypto";
-
-export async function getPoseidon(num: number): Promise<BaseContract> {
+export async function getPoseidon(ethers: HardhatEthers, num: number): Promise<BaseContract> {
   if (num < 1 || num > 6) {
     throw new Error("Poseidon Hash: Invalid number");
   }
 
   const [deployer] = await ethers.getSigners();
-  const PoseidonHasher = new ethers.ContractFactory(
+  const PoseidonHasher = new ContractFactory(
     poseidonContract.generateABI(num),
     poseidonContract.createCode(num),
     deployer,
@@ -25,12 +24,12 @@ export async function getPoseidon(num: number): Promise<BaseContract> {
 }
 
 export function poseidonHash(data: string): string {
-  data = ethers.hexlify(data);
+  data = hexlify(data);
 
   const chunks = splitHexIntoChunks(data.replace("0x", ""), 64);
   const inputs = chunks.map((v) => BigInt(v));
 
-  return ethers.toBeHex(Poseidon.hash(inputs), 32);
+  return toBeHex(Poseidon.hash(inputs), 32);
 }
 
 function splitHexIntoChunks(hexString: string, chunkSize = 64) {