+++ test/libs/arrays/Paginator.test.ts
@@ -1,12 +1,18 @@
-import { ethers } from "hardhat";
 import { expect } from "chai";
-import { Reverter } from "@/test/helpers/reverter";
+import hre from "hardhat";
 
-import { PaginatorMock } from "@ethers-v6";
 import { BigNumberish } from "ethers";
 
+import { HardhatEthers } from "@nomicfoundation/hardhat-ethers/types";
+
+import { Reverter } from "@test-helpers";
+
+import { PaginatorMock } from "@ethers-v6";
+
+const { ethers, networkHelpers } = await hre.network.connect();
+
 describe("Paginator", () => {
-  const reverter = new Reverter();
+  const reverter: Reverter = new Reverter(networkHelpers);
 
   let mock: PaginatorMock;