+++ test/libs/bn/U512.test.ts
@@ -1,11 +1,14 @@
-import { ethers } from "hardhat";
 import { expect } from "chai";
-import { Reverter } from "@/test/helpers/reverter";
+import hre from "hardhat";
+
+import { Reverter } from "@test-helpers";
 
 import { U512Mock } from "@ethers-v6";
 
+const { ethers, networkHelpers } = await hre.network.connect();
+
 describe("U512", () => {
-  const reverter = new Reverter();
+  const reverter: Reverter = new Reverter(networkHelpers);
 
   const prime = 76884956397045344220809746629001649092737531784414529538755519063063536359079n;