+++ test/libs/crypto/ECDSA384.test.ts
@@ -1,8 +1,11 @@
-import { ethers } from "hardhat";
 import { expect } from "chai";
-import { Reverter } from "@/test/helpers/reverter";
+import hre from "hardhat";
+
+import { Reverter } from "@test-helpers";
 
 import { ECDSA384Mock } from "@ethers-v6";
+
+const { ethers, networkHelpers } = await hre.network.connect();
 
 function modifyLeft(value: string, modifier: string): string {
   let newSignature = "0x";
@@ -31,7 +34,7 @@ function modifyRight(value: string, modifier: string): string {
 }
 
 describe.skip("ECDSA384", () => {
-  const reverter = new Reverter();
+  const reverter: Reverter = new Reverter(networkHelpers);
 
   let ecdsa384: ECDSA384Mock;
 
@@ -108,12 +111,12 @@ describe.skip("ECDSA384", () => {
         const wrongSig =
           "0x3066023100a2fcd465ab5b507fc55941c1c6cd8286de04b83c94c6be25b5bdf58e27d86c3759d5f94ffcbd009618b6371bc51994f0023100d708d5045caa4a61cad42622c14bfb3343a5a9dc8fdbd19ce46b9e24c2aff84ba5114bb543fc4b0099f369079302b721";
 
-        await expect(ecdsa384.verifySECP384r1(message, wrongSig, pubKey)).to.be.reverted;
+        await expect(ecdsa384.verifySECP384r1(message, wrongSig, pubKey)).to.be.revert(ethers);
 
         const wrongPubKey =
           "0x3076301006072a8648ce3d020106052b81040022036200041d77728fada41a8a7a23fe922e4e2dc8881a94b72a0612077ad80eeef13ff3bbea92aeef36a0f65885417aea104b86b76aedc226e260f7d0eeea8405b9269f354d929e5a98cab64fe192db94ed9335b7395e38e99b8bfaf32effa163a92889f9";
 
-        await expect(ecdsa384.verifySECP384r1(message, signature, wrongPubKey)).to.be.reverted;
+        await expect(ecdsa384.verifySECP384r1(message, signature, wrongPubKey)).to.be.revert(ethers);
       });
 
       it("should not revert when message is hashed using SHA-384", async () => {